SQL Injection
This assignment presents code that creates an in-memory database and runs a number of queries with and without SQL injection, always dumping the results to the console. You need to modify the run_query() function to detect a potential SQL injection attack, prevent it, and display that information to the console. The specific type of SQL injection attack is the “OR value=value;” attack, and you must defend against that specific attack with value being any allowed SQL literal (numeric or quoted string).