Buffer Overflow
This assignment presents code that collects a number from the user and then displays the number and a secret account number. Entering too many characters triggers the overwrite of the account number. Using the existing source code as a starting point, you need to prevent the user input from impacting the account number.